+++ packages/daemon/src/app/controller/pipeline.ts
@@ -4,7 +4,7 @@ import * as HttpStatus from 'http-status';
 import * as Joi from 'joi';
 import Debug from 'debug';
 import { PluginManager } from '../../service/plugin';
-import { parseConfig } from '../../runner/helper';
+import { parseConfig } from '../../utils';
 import { BaseEventController } from './base';
 import { PipelineService } from '../../service/pipeline';
 import { Tracer } from '../../service/trace-manager';
